requirements.txt
pillow
colorama
requests
tk
aiohttp